shm/requirements.txt
flask
mysql-connector-python
requests
fastapi
uvicorn